requirements.txt
Pillow
pyperclip
requests
